+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.5.1
 
+  - rename few static functions
+
 ## 2.5.0
 
   - add functions :