+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.5.1
 
   - rename few static functions
+  - deprecated mak mame and raine (use run:mame and run:raine)
 
 ## 2.5.0